+++ app/src/main/java/com/fallgamlet/dnestrcinema/mvp/models/AccountItem.java
@@ -13,7 +13,7 @@ public class AccountItem {
     @SerializedName("password")
     private String password;
     @SerializedName("cinemaId")
-    private int cinemaId;
+    private long cinemaId;
 
 
     public AccountItem() {
@@ -39,11 +39,11 @@ public class AccountItem {
         this.password = password;
     }
 
-    public int getCinemaId() {
+    public long getCinemaId() {
         return cinemaId;
     }
 
-    public void setCinemaId(int cinemaId) {
+    public void setCinemaId(long cinemaId) {
         this.cinemaId = cinemaId;
     }
 }
